refactor: column type changed to a more proper for column position values

+++ app/Domain/Member/Database/Migrations/2025_10_01_201912_create_members_moderation_type_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('key', 64)->unique();
             $table->string('title', 64);
             $table->string('description', 256);
-            $table->tinyInteger('position')->default('0');
+            $table->smallInteger('position')->default('0');
             $table->timestamps();
         });
     }
